step 1

+++ src/service/product.js
@@ -16,6 +16,11 @@ export default {
   },
 
   async save(form) {
+    if (store.get('session/isEditor') != true) {
+      store.dispatch('warning/youMustBeEditor');
+      return;
+    }
+
     let method = 'post', opType = 'added';
     if (form.id && form.id > 0) {
       method = 'put'; opType = 'edited';
@@ -39,11 +44,21 @@ export default {
   },
 
   async toggle(id) {
+    if (store.get('session/isEditor') != true) {
+      store.dispatch('warning/youMustBeEditor');
+      return;
+    }
+
     const res = await Helper.call('Product Toggle', { method: 'put', url: baseURL + '/toggle/' + id });
     return res.status;
   },
 
   async remove(id) {
+    if (store.get('session/isEditor') != true) {
+      store.dispatch('warning/youMustBeEditor');
+      return;
+    }
+
     const res = await Helper.call('Product Delete', { method: 'delete', url: baseURL + '/' + id });
     return res.status;
   }
